Fix mistake in READ_PTR macros, use sizeof(void*)

+++ include/global.h
@@ -171,14 +171,14 @@ int strcmp(const char *, const char*);
 #define T1_READ_16(ptr) ((ptr)[0] | ((ptr)[1] << 8))
 #define T1_READ_32(ptr) ((ptr)[0] | ((ptr)[1] << 8) | ((ptr)[2] << 16) | ((ptr)[3] << 24))
 #define T1_READ_64(ptr) ((ptr)[0] | ((ptr)[1] << 8) | ((ptr)[2] << 16) | ((ptr)[3] << 24) | ((ptr)[4] << 32) | ((ptr)[5] << 40) | ((ptr)[6] << 48) | ((ptr)[7] << 56))
-#define T1_READ_PTR(ptr) (u8 *) T1_READ_64((u64*)ptr)
+#define T1_READ_PTR(ptr) (u8 *) T1_READ_64((u8*)ptr)
 
 // T2_READ_8 is a duplicate to remain consistent with each group.
 #define T2_READ_8(ptr)  ((ptr)[0])
 #define T2_READ_16(ptr) ((ptr)[0] + ((ptr)[1] << 8))
 #define T2_READ_32(ptr) ((ptr)[0] + ((ptr)[1] << 8) + ((ptr)[2] << 16) + ((ptr)[3] << 24))
 #define T2_READ_64(ptr) ((ptr)[0] + ((ptr)[1] << 8) + ((ptr)[2] << 16) + ((ptr)[3] << 24) + ((ptr)[4] << 32) + ((ptr)[5] << 40) + ((ptr)[6] << 48) + ((ptr)[7] << 56) )
-#define T2_READ_PTR(ptr) (void *) T2_READ_64((u64*)ptr)
+#define T2_READ_PTR(ptr) (void *) T2_READ_64((u8*)ptr)
 
 // Macros for checking the joypad
 #define TEST_BUTTON(field, button) ((field) & (button))
